Change totalShards type to long

+++ src/main/kotlin/net/trueog/questsOG/quests/HomesFive.kt
@@ -15,7 +15,7 @@ import org.bukkit.persistence.PersistentDataType
 
 class HomesFive : Quest {
     private data class Requirements(
-        val totalShards: Int,
+        val totalShards: Long,
         val ticksPlayed: Int,
         val pigOneCm: Int,
         val striderOneCm: Int,
