+++ src/main/kotlin/net/trueog/questsOG/quests/HomesTwo.kt
@@ -10,7 +10,7 @@ import org.bukkit.entity.Player
 
 class HomesTwo : Quest {
     private data class Requirements(
-        val totalShards: Int,
+        val totalShards: Long,
         val ticksPlayed: Int,
         val totalCm: Int,
         val levels: Int,
